+++ s/index.ts
@@ -1,4 +1,5 @@
 
 export * from "./layout/index.js"
 export * from "./studio/index.js"
+export * from "./integrations/index.js"